# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 이어봄 Service Metadata API
  version: 1.0.0-target-draft
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    서비스 계정, 구독, 가정, 가족 초대와 불투명 로컬 프로필 연결만 처리한다.
    건강기록, 가족력, 통증 기록, 건강서류, OCR 결과, 예측 결과와 로컬 변경 이력은
    이 API의 요청·응답·로그·작업 큐에 포함할 수 없다.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /family-invitations:
    post:
      tags: [Invitations]
      operationId: createFamilyInvitation
      summary: 기존 가족 구성원 로컬 프로필을 대상으로 서비스 계정 초대를 생성한다.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvitationCreateRequest' }
      responses:
        '201':
          description: 초대 생성 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationResponse' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Invitations]
      operationId: listFamilyInvitations
      summary: 현재 계정이 보냈거나 받은 초대를 조회한다.
      parameters:
        - name: direction
          in: query
          required: false
          schema: { type: string, enum: [sent, received], default: received }
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/InvitationStatus' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 초대 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationPage' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /family-invitations/{invitation_id}:
    parameters:
      - $ref: '#/components/parameters/InvitationId'
    get:
      tags: [Invitations]
      operationId: getFamilyInvitation
      summary: 초대 상세 상태를 조회한다.
      responses:
        '200':
          description: 조회 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Invitations]
      operationId: cancelFamilyInvitation
      summary: 보낸 대기 중 초대를 취소한다.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: 취소 완료
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /family-invitations/{invitation_id}/accept:
    post:
      tags: [Invitations]
      operationId: acceptFamilyInvitation
      summary: 받은 초대를 수락한다.
      description: 초대 수락은 계정 참여만 확정하며 건강정보 다운로드나 로컬 프로필 자동 생성은 실행하지 않는다.
      parameters:
        - $ref: '#/components/parameters/InvitationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: 수락 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationResponse' }
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          description: 초대 만료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '412':
          $ref: '#/components/responses/PreconditionFailed'
  /family-invitations/{invitation_id}/decline:
    post:
      tags: [Invitations]
      operationId: declineFamilyInvitation
      summary: 받은 초대를 거절한다.
      parameters:
        - $ref: '#/components/parameters/InvitationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: 거절 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationResponse' }
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    InvitationId:
      name: invitation_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 동일 계정·operation에서 24시간 동안 유효한 16~72자의 클라이언트 생성 키
      schema:
        type: string
        minLength: 16
        maxLength: 72
        pattern: '^[A-Za-z0-9._:-]+$'
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: '조회 응답의 row_version. 예: "3"'
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 512 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: Access Token 누락·만료·위조
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 인증됐으나 대상 작업 권한 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 대상이 없거나 조회 권한이 없어 존재를 숨김
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 현재 상태 또는 유일성 제약과 충돌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PreconditionFailed:
      description: If-Match row_version 불일치
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: 요청 스키마 또는 비즈니스 입력 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    InvitationStatus:
      type: string
      enum: [pending, accepted, declined, expired, cancelled]
    OpaqueLocalRef:
      type: string
      description: 한 연결 생명주기에만 사용하는 32바이트 이상 CSPRNG base64url 불투명 값. 같은 작업의 재시도에서는 유지하고 종료 후에는 재사용하지 않으며 프로필 내용을 인코딩하지 않는다.
      minLength: 43
      maxLength: 86
      pattern: '^[A-Za-z0-9_-]+$'
    InvitationCreateRequest:
      type: object
      additionalProperties: false
      required: [household_id, invitee_email, target_profile_ref]
      properties:
        household_id: { type: string, format: uuid }
        invitee_email: { type: string, format: email, maxLength: 320 }
        target_profile_ref: { $ref: '#/components/schemas/OpaqueLocalRef' }
    InvitationResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - household_id
        - direction
        - counterparty_email
        - target_profile_ref
        - status
        - expires_at
        - created_at
        - updated_at
        - row_version
      properties:
        id: { type: string, format: uuid }
        household_id: { type: string, format: uuid }
        direction: { type: string, enum: [sent, received] }
        counterparty_email: { type: string, format: email }
        target_profile_ref: { $ref: '#/components/schemas/OpaqueLocalRef' }
        status: { $ref: '#/components/schemas/InvitationStatus' }
        expires_at: { type: string, format: date-time }
        accepted_at: { type: [string, 'null'], format: date-time }
        declined_at: { type: [string, 'null'], format: date-time }
        cancelled_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        row_version: { type: integer, format: int64, minimum: 1 }
    InvitationPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/InvitationResponse' }
        next_cursor: { type: [string, 'null'] }
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field: { type: string, example: invitee_email }
        reason: { type: string, example: invalid_format }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
            message: { type: string }
            request_id: { type: string, format: uuid }
            field_errors:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
            details:
              type: object
              additionalProperties: true